comit tower+jump edge coll

+++ Assets/Scripts/Player/MovementBehaviours/MovementBehaviour.cs
@@ -110,7 +110,7 @@ public abstract class MovementBehaviour : MonoBehaviour {
         Grounded = false;
         RaycastHit2D hit = Physics2D.Raycast(m_GroundCheck.position, Vector2.down, 10, m_FloorMask);
 		Debug.Log (hit.distance);
-		if (hit && hit.distance == 0)   
+		if (hit && (int)hit.distance == 0)   
         {
             Grounded = true;
